+++ _result/golang/Aufgabe.go
@@ -1,6 +1,6 @@
 package golang
 
 type Aufgabe struct {
-	AufgabeNr   float64 `json:AufgabeNr`
-	Bezeichnung string  `json:Bezeichnung`
+	AufgabeNr   int    `json:AufgabeNr`
+	Bezeichnung string `json:Bezeichnung`
 }
