+++ _result/golang/Buchung-Flat.go
@@ -1,11 +1,11 @@
 package golang
 
 type BuchungFlat struct {
-	BuchungNr                      float64      `json:BuchungNr`
-	BuchungszeileNr                float64      `json:BuchungszeileNr`
+	BuchungNr                      int          `json:BuchungNr`
+	BuchungszeileNr                int          `json:BuchungszeileNr`
 	Buchungsdatum                  string       `json:Buchungsdatum`
 	Belegdatum                     string       `json:Belegdatum`
-	Belegnummer                    float64      `json:Belegnummer`
+	Belegnummer                    int          `json:Belegnummer`
 	Kondition                      Kondition    `json:Kondition`
 	Buchungsart                    Buchungsart  `json:Buchungsart`
 	EinzahlungMitteilung           string       `json:EinzahlungMitteilung`
@@ -14,16 +14,16 @@ type BuchungFlat struct {
 	Zahlungsart                    Zahlungsart  `json:Zahlungsart`
 	EsrKodierzeile                 string       `json:EsrKodierzeile`
 	EsrPruefziffer                 string       `json:EsrPruefziffer`
-	Herkunft                       float64      `json:Herkunft`
+	Herkunft                       int          `json:Herkunft`
 	Auftrag                        Auftrag      `json:Auftrag`
 	Belegart                       Belegart     `json:Belegart`
-	EsrNummer                      float64      `json:EsrNummer`
+	EsrNummer                      int          `json:EsrNummer`
 	EinzahlungName                 string       `json:EinzahlungName`
 	HabenKonto                     Konto        `json:HabenKonto`
 	SollKonto                      Konto        `json:SollKonto`
 	Waehrung                       Waehrung     `json:Waehrung`
 	Kurs                           float64      `json:Kurs`
-	MahnCode                       float64      `json:MahnCode`
+	MahnCode                       int          `json:MahnCode`
 	MahnDatum                      string       `json:MahnDatum`
 	Steuercode                     Steuercode   `json:Steuercode`
 	SteuerbetragFW                 float64      `json:SteuerbetragFW`
