+++ _result/golang/Funktion.go
@@ -1,6 +1,6 @@
 package golang
 
 type Funktion struct {
-	FunktionNr  float64 `json:FunktionNr`
-	Bezeichnung string  `json:Bezeichnung`
+	FunktionNr  int    `json:FunktionNr`
+	Bezeichnung string `json:Bezeichnung`
 }
