+++ _result/golang/Gutschein.go
@@ -3,10 +3,10 @@ package golang
 type Gutschein struct {
 	GutscheinNr      string      `json:GutscheinNr`
 	Bezeichnung      string      `json:Bezeichnung`
-	Anzahl           float64     `json:Anzahl`
-	AnzahlEingeloest float64     `json:AnzahlEingeloest`
-	AnzahlProAdresse float64     `json:AnzahlProAdresse`
-	AnzahlVerkauft   float64     `json:AnzahlVerkauft`
+	Anzahl           int         `json:Anzahl`
+	AnzahlEingeloest int         `json:AnzahlEingeloest`
+	AnzahlProAdresse int         `json:AnzahlProAdresse`
+	AnzahlVerkauft   int         `json:AnzahlVerkauft`
 	Betrag           float64     `json:Betrag`
 	Buchungsart      Buchungsart `json:Buchungsart`
 	GueltigBis       string      `json:GueltigBis`
