+++ _result/golang/Kondition.go
@@ -1,15 +1,15 @@
 package golang
 
 type Kondition struct {
-	KonditionNr    float64 `json:KonditionNr`
+	KonditionNr    int     `json:KonditionNr`
 	Bezeichnung    string  `json:Bezeichnung`
 	Inaktiv        bool    `json:Inaktiv`
 	EndeJahr       bool    `json:EndeJahr`
 	EndeMonat      bool    `json:EndeMonat`
 	EndeQuartal    bool    `json:EndeQuartal`
-	NettoTage      float64 `json:NettoTage`
-	Skonto1Tage    float64 `json:Skonto1Tage`
+	NettoTage      int     `json:NettoTage`
+	Skonto1Tage    int     `json:Skonto1Tage`
 	Skonto1Prozent float64 `json:Skonto1Prozent`
-	Skonto2Tage    float64 `json:Skonto2Tage`
+	Skonto2Tage    int     `json:Skonto2Tage`
 	Skonto2Prozent float64 `json:Skonto2Prozent`
 }
