+++ _result/golang/OffenerRapport.go
@@ -1,7 +1,7 @@
 package golang
 
 type OffenerRapport struct {
-	OffenerRapportNr          float64        `json:OffenerRapportNr`
+	OffenerRapportNr          int            `json:OffenerRapportNr`
 	Mitarbeiter               Mitarbeiter    `json:Mitarbeiter`
 	Auftrag                   Auftrag        `json:Auftrag`
 	Bezeichnung               string         `json:Bezeichnung`
@@ -9,8 +9,8 @@ type OffenerRapport struct {
 	Bemerkungen               string         `json:Bemerkungen`
 	BemerkungenRTF            string         `json:BemerkungenRTF`
 	Serviceauftrag            Serviceauftrag `json:Serviceauftrag`
-	ServiceauftragspositionNr float64        `json:ServiceauftragspositionNr`
-	DauerTotal                float64        `json:DauerTotal`
+	ServiceauftragspositionNr int            `json:ServiceauftragspositionNr`
+	DauerTotal                int            `json:DauerTotal`
 	Kontaktname               string         `json:Kontaktname`
 	Kontakt                   Kontakt        `json:Kontakt`
 	Rapport                   Rapport        `json:Rapport`
@@ -22,5 +22,5 @@ type OffenerRapport struct {
 	StartZeit                 string         `json:StartZeit`
 	EndDatum                  string         `json:EndDatum`
 	EndZeit                   string         `json:EndZeit`
-	Dauer                     float64        `json:Dauer`
+	Dauer                     int            `json:Dauer`
 }
