+++ _result/golang/Preis.go
@@ -3,7 +3,7 @@ package golang
 type Preis struct {
 	Bezeichnung   string       `json:Bezeichnung`
 	Preisstaffel  Preisstaffel `json:Preisstaffel`
-	Preiscode     float64      `json:Preiscode`
+	Preiscode     int          `json:Preiscode`
 	StaffelVon    float64      `json:StaffelVon`
 	StaffelBis    float64      `json:StaffelBis`
 	Wert          float64      `json:Wert`
