+++ _result/golang/Teilleistungsgruppe.go
@@ -1,7 +1,7 @@
 package golang
 
 type Teilleistungsgruppe struct {
-	TeilleistungsgruppeNr float64 `json:TeilleistungsgruppeNr`
-	Bezeichnung           string  `json:Bezeichnung`
-	Gliederung            string  `json:Gliederung`
+	TeilleistungsgruppeNr int    `json:TeilleistungsgruppeNr`
+	Bezeichnung           string `json:Bezeichnung`
+	Gliederung            string `json:Gliederung`
 }
