+++ _result/golang/Teilleistungsuntergruppe.go
@@ -1,7 +1,7 @@
 package golang
 
 type Teilleistungsuntergruppe struct {
-	TeilleistungsuntergruppeNr float64 `json:TeilleistungsuntergruppeNr`
-	Bezeichnung                string  `json:Bezeichnung`
-	Gliederung                 string  `json:Gliederung`
+	TeilleistungsuntergruppeNr int    `json:TeilleistungsuntergruppeNr`
+	Bezeichnung                string `json:Bezeichnung`
+	Gliederung                 string `json:Gliederung`
 }
