+++ _result/golang/Zivilstand.go
@@ -1,6 +1,6 @@
 package golang
 
 type Zivilstand struct {
-	ZivilstandNr float64 `json:ZivilstandNr`
-	Bezeichnung  string  `json:Bezeichnung`
+	ZivilstandNr int    `json:ZivilstandNr`
+	Bezeichnung  string `json:Bezeichnung`
 }
